requirements.txt
# DeepEval API Wrapper - Essential Dependencies
# Optimized for Render deployment

# Core FastAPI stack
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# DeepEval and core dependencies
deepeval==3.4.1

# Basic utilities
requests>=2.31.0